% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_wrapper.R
\name{da_metrics}
\alias{da_metrics}
\title{FDR and Power for Differential Analysis}
\usage{
da_metrics(results, nonnull, level = 0.05, focus_col = ncol(results))
}
\description{
FDR and Power for Differential Analysis
}
