% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_wrapper.R
\name{run_ancombc2}
\alias{run_ancombc2}
\title{This function run ANCOM-BC2 method on a TreeSummarizedExperiment object.}
\usage{
run_ancombc2(tseObj, assay_name = "counts", fixed_formula, p_adj_method = "BH")
}
\arguments{
\item{tseObj}{the input data. The data parameter should be either a phyloseq or a TreeSummarizedExperiment object,
which consists of a feature table (microbial count table), a sample metadata table,
a taxonomy table (optional), and a phylogenetic tree (optional).}

\item{assay_name}{character. Default is "counts". See ?SummarizedExperiment::assay for more details.}

\item{fixed_formula}{the character string expresses how the microbial absolute abundances for each taxon depend on the fixed effects in metadata.}

\item{p_adj_method}{the multiple correction procedure to be used.}
}
\value{
a data frame with log fold changes, p_values, q_valuesh.
}
\description{
This function run ANCOM-BC2 method on a TreeSummarizedExperiment object.
}
