% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_wrapper.R
\name{run_DA_method}
\alias{run_DA_method}
\title{Wrapper function for each DA_method}
\usage{
run_DA_method(raw_data, meta_data, DA_method)
}
\arguments{
\item{raw_data:}{an otu table with Otus on rows and Samples/Subjects on columns.}

\item{meta_data:}{meta data containing group information and other covariate information on subjects}

\item{DA_method:}{character, the DA method.}
}
\value{
The output of the DA method (log fold change across groups, pvalue, adj.pvalue)
}
\description{
Wrapper function for each DA_method
}
