% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_wrapper.R
\name{run_DESeq}
\alias{run_DESeq}
\title{This function run DESeq method on a DESeq object.}
\usage{
run_DESeq(
  dds,
  sftype = "poscounts",
  tidy = TRUE,
  format = "DataFrame",
  pAdjustMethod = "BH"
)
}
\arguments{
\item{dds}{a DESeqDataSet object.}

\item{sftype}{either "ratio", "poscounts", or "iterate" for the type of size factor estimation.}

\item{tidy}{whether to output the results table with rownames as a first column}

\item{format}{character, either "DataFrame", "GRanges", or "GRangesList"}

\item{pAdjustMethod}{the multiple correction procedure to be used.}
}
\value{
a data frame with log fold changes, p_values, q_valuesh.
}
\description{
This function run DESeq method on a DESeq object.
}
