% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_wrapper.R
\name{run_limma}
\alias{run_limma}
\title{This function run LIMMA-VOOM method on a TreeSummarizedExperiment object.}
\usage{
run_limma(dge, metaD, formula, p_adj_method = "BH")
}
\arguments{
\item{dge}{a numeric matrix containing raw counts, or an ExpressionSet containing raw counts, or a DGEList object.}

\item{metaD}{meta data containing group information and other covariate information on subjects}

\item{formula}{the character string expresses how the microbial absolute abundances for each taxon depend on the fixed effects in metadata.}

\item{p_adj_method}{the multiple correction procedure to be used.}
}
\value{
a data frame with log fold changes, p_values, q_valuesh.
}
\description{
This function run LIMMA-VOOM method on a TreeSummarizedExperiment object.
}
