% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{spline_fun}
\alias{spline_fun}
\title{Generate Random Spline}
\usage{
spline_fun(D = 2, knots = NULL, h_ix = 1:10, ...)
}
\description{
This generates random spline functions. It is used in the toy nonlinear
dataset created in \code{demo_spline()}. This is not necessary for the essential
multimedia workflow, it is only exported for potential independent interest.
}
\examples{
x <- seq(-2, 2, length.out = 100)
f <- spline_fun(sd = 0.3)
fx <- f(x)
plot(x, fx[, 1])
}
\seealso{
demo_spline
}
