% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{splsda_fit}
\alias{splsda_fit}
\title{Run SPLS-DA and Collect Metrics}
\usage{
splsda_fit(
  exper,
  outcome_var = "outcome2",
  ncomp = 2,
  keepX = NULL,
  nrepeat = 4,
  folds = 5
)
}
\value{
A list with both the CV-averaged AUC and the fitted SPLS-DA object.
}
\description{
Wrapper to apply mixOmics' sparse pls discriminant analysis ' on a
SummarizedExperiment object.
}
