% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{splsda_fit}
\alias{splsda_fit}
\title{Run SPLS-DA and Collect Metrics}
\usage{
splsda_fit(
  exper,
  outcome_var = "outcome2",
  ncomp = 2,
  keepX = NULL,
  nrepeat = 4,
  folds = 5
)
}
\arguments{
\item{exper}{The SummarizedExperiment on which to run spls-da.}

\item{outcome_var}{The response variable from exper's colData slot.}

\item{ncomp}{The number of components to use in the SPLS-DA.}

\item{keepX}{The number of variables to include in each component.}

\item{folds}{The number of folds used for evaluating performance with CV.}

\item{repeat}{The number of repeats used for evaluating performance with CV.}
}
\value{
A list with both the CV-averaged AUC and the fitted SPLS-DA object.
}
\description{
Wrapper to apply mixOmics' sparse pls discriminant analysis ' on a
SummarizedExperiment object.
}
